Speed
Added a fast preview mode. Will be limited to grayscale, but is signifigantly faster, regardless of chunk size.

Added a function to directly read tags from nested compounds in binary. Signifigantly faster than the conversion to DS scripts, but doesn't support list entries yet (Grabs whole list), and of course, can't be exported to json or used to visualize an NBT struct in debugger.

Added warning for old change (not important)