Added GUI Exporting
Press 'E' to export a selection.
Added an object for each chunk, with a modified version of the export script so that it exports as many blocks as it can in a specific amount of  time and then stops (Not section or chunk based, just each block).

Time is a set amount, divided between all of the selected chunks.